+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed:
 - Replace modal_bottom_sheet lib with Flutter's own implementation
+- Replace badges lib with Flutter's own implementation
 
 ## [0.8.268] - 2023-02-19
 ### Fixed: